#17517: Font rendering different after PDF export
git-svn-id: svn://scribus.net/branches/Version16x/Scribus@26854 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/pdflib_core.cpp
@@ -165,10 +165,10 @@ public:
 
 	bool       firstUse { true };
 	QByteArray fontName;
-	int        fontSize { -1 };
+	double     fontSize { -1.0 };
 	QByteArray fillColor;
 	QByteArray strokeColor;
-	int        strokeWidth { -1 };
+	double     strokeWidth { -1.0 };
 	int        renderingMode { 0 };
 	QTransform textMatrix;
 };
